+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
-## [27.1.0-next.0](https://github.com/eduflow/ckeditor5-iframe/compare/v27.1.0...v27.1.0) (current)
+## [27.1.0-next.1](https://github.com/eduflow/ckeditor5-iframe/compare/v27.1.0-next.0...v27.1.0-next.1)
 
--   Initial
+- Rename `HTMLEmbed` -> `IframeEmbed`, `htmlEmbed` -> `iframeEmbed`
+- Add icon courtesy of [Material Design
+  Icons](https://iconify.design/icon-sets/mdi/iframe.html)
+
+## [27.1.0-next.0](https://github.com/eduflow/ckeditor5-iframe/tree/v27.1.0-next.0)
+
+- Initial import based on [HTML Embed](https://ckeditor.com/docs/ckeditor5/latest/features/html-embed.html)